deadsands/pyproject.toml
[tool.poetry]
name = 'dnd'
version = '1.0'
license = 'The Unlicense'
authors = ['Greg Boyington <evilchili@gmail.com>']
description = 'Source for a homebrew DND campaign toolset.'
packages = [
    { include = "site_tools" }
]

[tool.poetry.dependencies]
python = "^3.10"

# cli business
typer = "latest"
rich = "latest"
pyyaml = "latest"
livereload = "latest"
jinja2-simple-tags = "latest"
pyinotify = "latest"

# static website
pelican = "^4.7.2"
pelican-drafts = "^0.1.1"
pelican-sitemap = "^1.0.2"
pelican-yaml-metadata = { git = "https://github.com/pR0Ps/pelican-yaml-metadata.git", branch = "cdc1b9708916410e455e8e258e3d39a9d575c7b5" }

# local wotsits
dnd-npcs      = { git = "https://github.com/evilchili/dnd-npcs", branch = 'main' }
elethis-cipher= { git = "https://github.com/evilchili/elethis-cipher", branch = 'main' }
#dnd-rolltable = { git = "https://github.com/evilchili/dnd-rolltable", branch = 'main' }
dnd-rolltable = { file = "../../dnd-rolltable/dist/dnd_rolltable-1.1.9-py3-none-any.whl" }
prompt-toolkit = "^3.0.38"

[tool.poetry.scripts]
site = "site_tools.cli:app"
roll-table = "rolltable.cli:app"
pelican = "site_tools.tasks:pelican_main"



[build-system]
requires = ['poetry-core~=1.0']
build-backend = 'poetry.core.masonry.api'
